fix multi-platform for linux
