+++ pkg/platform/other.go
@@ -1,0 +1,5 @@
+// +build !darwin
+
+package platform
+
+func start(h *CursorHandle) {}
